update changePassword location
